elasticsearch/requirements.txt
elasticsearch>=8,<9
psycopg
sshtunnel
python-dotenv
